+++ tests/unit.test.ts
@@ -395,3 +395,20 @@ describe("Rotation", () => {
 		expect(round(degrees.getTheta(true), 2)).toBe(15);
 	});
 });
+
+describe("Representation", () => {
+	it("Returns vector's components packed into array", () => {
+		const fn = (input: Input) => vector(input).toArray();
+
+		expect(fn({ x: 1 })).toEqual([ 1, 0, 0 ]);
+		expect(fn({ x: 1, y: 2 })).toEqual([ 1, 2, 0 ]);
+		expect(fn({ x: 1, y: 2, z: 3 })).toEqual([ 1, 2, 3 ]);
+	});
+	it("Converts an instance into primitive value", () => {
+		const a = vector([ 3, 4 ]);
+		const b = vector([ 6, 8 ]);
+
+		// @ts-expect-error: testing coercion
+		expect(a + b).toBe(15);
+	});
+});